Updates to utilize new App Router structure
- Converts to typescript
- Removes no longer needed Page Router dir
- Sets up to use new App Router structure
