+++ app/(routes)/musica/page.tsx
@@ -1,0 +1,16 @@
+import Head from 'next/head'
+import Layout from '@/components/layout'
+import Content from '@/components/grid/container'
+
+export default function Home() {
+  return (
+    <Layout>     
+      <Content>
+        <h1>
+          Musica
+        </h1>
+
+      </Content>
+    </Layout>
+  )
+}
