+++ app/(routes)/trabajo/page.tsx
@@ -1,0 +1,9 @@
+import Container from '@/components/grid/container'
+
+export default function Trabajo() {
+  return (
+    <Container className="">
+      <h1>Trabajo</h1>
+    </Container>
+  )
+}
