--- app/components/layout.js
@@ -1,5 +1,0 @@
-export default function Layout({ children }) {
-  return <div className='flex flex-col md:flex-row justify-between md:justify-start h-screen'>
-    {children}
-  </div>
-}
