--- app/pages/_app.js
@@ -1,7 +1,0 @@
-import '../styles/globals.scss'
-
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
